docs: update roadmap and improve consistency (#670)
The roadmap on the docs site is out of date, and the roadmap on the main
page is inconsistent with the roadmap page. I made the two lists
consistent and ensured they represent the current state of our progress.
I also added some links to relevant areas of the docs.