fix(svm): conform on some more crypto syscalls (#709)
The first commit is unfortunately pretty nasty. It seems that Agave/FD
doesn't check the security of the precomputed element when doing a
scalar multiplication over Edwards25519. The Zig stdlib provides this
check by default without providing a way to remove it (for very good
reason), but in the end, we need to match Agave, so I've pulled out the
necessary functions and commented out the checks.