refactor(consensus, tower): Clearer Tower API (#728)
Supersedes #724 

Moved to a separate PR to be able to merge directly without being
blocked by the previous upstream PR.

This refactor PR mostly moves things around. 

The objective is to simplify the Tower API and make it more explicit. 

Methods were extracted into a new structure (currently named
`ReplayTower` - other suggestion for the new name is `Consensus`) which
houses the extra methods that is used during Replay. Some of these
methods depends on the extracted tower, while some don't.

---------

Co-authored-by: Drew Nutter <dnut@users.noreply.github.com>