feat(replay): scaffolding for replay stage (#726)
This just sets up some high-level scaffolding for replay without any
actual logic implemented. This is a minimal PR so dadepo and I can have
a shared space to start working in asap. Most of the logic I've
implemented for replay is in separate branches which can be reviewed and
merged later.

Changes:
- Add `replay` folder with:
  - `service.zig`: The high level state and loop for executing replay.
- `trackers.zig`: SlotTracker and EpochTracker which are used for now to
keep the state in agave's bankforks. I'd like to avoid passing these
structs around to too many places, but we need to keep this state in
replay at a high level
- run replay in the `validator` command in cmd.zig
- Add `Logger.from` for more ergonomic switching of scopes (this will
come in handy in 0.14 with decl literals)