feat(replay): lock accounts and resolve lookup table addresses (#769)
`resolve_lookup.zig`: Replay needs to convert transactions into a
"resolved" format where all the addresses are pulled out of the lookup
tables from accountsdb and the data structures are reorganized for easy
consumption for the svm and the account locks

`account_locks.zig`: Replay needs to enforce account locking. If an
account is written to in a single transaction within a batch, no other
transactions may touch that account.