fix(collections): sorted map - preserve sorting in bincode (#799)
# Problem

If you serialize a SortedMap and then deserialize it, it becomes
unsorted. See the test case I added. This test fails on main. This
breaks the ledger database which is a problem for replaying blocks.


# Cause

### Background

SortedMap is sorted lazily. Write operations do not result in a sorted
map. The map is only sorted before read operations. It uses the `max`
and `is_sorted` fields to determine whether or not the map needs to be
sorted before the next read.

### Problem

A bincode config was added in #784 to exclude the `max` and `is_rooted`
fields. This fields are being set incorrectly after deserialization.
Likewise, unsorted data is assumed to be sorted.

# Fix

I removed the bincode config.