fix PIE compilation with TSan and remove `codecov.yml` (#809)
As you can see here, https://docs.codecov.com/docs/github-checks,
CodeCov has deprecated its annotations feature. This PR removes the
`codecov.yml` file that was disabling it and bumps the orb version to
one that outputs a config file that doesn't contain the `annotations`
field, so that CodeCov interprets it as valid.