chore(build, log): make some fixes and cleanup (#771)
- Moved away from using the `comptime {` block to reference tests, since
this ends up pulling in all of the tests when compiling other stuffs and
increases the compile times needlessly.
- Added a `check` step so we can use incremental compilation with ZLS
smoother
- Updates some more of the package hashes to make sure they're using the
updated package hash themselves. ZLS will try to build the dependency as
the root, which will fail on 0.14 unless it uses the newer package hash.
Also improves the incremental experience.
- Removed an extremely silly function that should never exist!