feat(svm): use account reader interface in svm's account loader
Also adds ForkAccountReader interface for typical pattern where ancestors are needed. Typically ancestors are passed around only for this use case. The user of the account reader usually only cares about the current fork it's processing. So it can just depend on an account reader that guarantees it will only handle one fork.