fix(accountsdb): fix rare OOM when loading snapshots (#818)
Before, code using an FBA for the bufferpool's temp allocations would be
slightly undersized, and would cause an OOM reading accounts accounts
very close (a few bytes) to the 10MiB max.

This is because this function (readAccount) calls getSlice twice with
the same allocator - the first read would use memory inside the FBA,
leading us to have slightly too little for the second read.

This fix just deallocates the data used for the first read earlier such
that the 2nd read has the memory it may need.