fix(svm,accountsdb): account loader fork-awareness and 0-lamport handling (#819)
This PR makes the svm's account loader fork-aware and fixes the handling
of 0-lamport accounts by switching the account loader to use one of the
new interfaces for accountsdb.

## Motivation

Replay will not work correctly unless the account loader uses fork-aware
(ancestors-based) accountsdb methods. If it's not fork-aware it will
provide incorrect account state into the SVM, resulting in incorrect
transaction outputs.

This also fixes a bug that's causing problems in the transaction harness
conformance tests. In the account loader, 0-lamport accounts are not
supposed to be included, but currently they are being returned by
accountsdb. The new interfaces handle that scenario properly. AccountsDB
should eventually be changed to do this, but it's tedious to make it
work in the most performant way. Fixing it through the interface is
sufficient for now.

Why switch to the new interface:
- New interfaces already include both fork-awareness and the 0-lamport
bugfix.
- Easier compatibility with replay, which uses these interfaces.
- Any other solution would be just as much work since it would require
reworking the existing comptime interface to add the functionality
supported by the new interfaces.
- I believe the new interfaces are also a better long-term approach.
Compared to the existing comptime interface, they are more readable and
more widely applicable to other areas of code, making it easier to
manage dependencies and interoperability of components.

## Details

`AccountStore` and `AccountReader` were added in #814 as interfaces for
`AccountsDB`. This PR also adds a new one called `SlotAccountReader`.

The git diff in this PR is a bit misleading about what code is actually
new. The `AccountStore` struct appears to be newly added in this PR
because I've moved it up from a lower location in the file. It's
otherwise identical. `SlotAccountReader` is the entirely new addition,
even though in the diff it appears to be an evolution of AccountStore.

The new `SlotAccountReader` interface is split off from `AccountReader`
to serve the typical use case where you're only concerned about the
state for one particular slot. Typically ancestors are passed around
only for this use case. Instead, you can simply depend on an account
reader that ensures it will only return accounts that should be visible
during that slot.

---------

Co-authored-by: Sobeston <15335529+Sobeston@users.noreply.github.com>
Co-authored-by: Harold Newman <hnewman2744@gmail.com>