fix(svm): copy before mutating nonce
previously these steps were run:
1. get pointer to nonce acct from cache
2. advance the nonce, mutating that pointer
3. copy the data behind the pointer into a rollback account

this is problematic because the first instruction in the transaction is supposed to be able to read the *old* version of the account so it can advance the nonce. but in this case, we're updating it before running the instructions, and the instruction reads the *already updated* version of the account, which leads to an unexpected error in the system program.

the solution here is to copy the rollback account *before* modifying the nonce. that way the account that will be read by the system program has the expected old state, and the rollback account has the advanced nonce (same as before, this is what we want). this is the same way that agave works, and it's convoluted, but Arc::make_mut clones the underlying vec (not the arc) which means you get a copy of the account