fix(svm): nonce account mutation timing & rent exemption check (#823)
# Nonce account mutation timing

previously these steps were run:
1. get pointer to nonce acct from cache
2. advance the nonce, mutating that pointer
3. copy the data behind the pointer into a rollback account

this is problematic because the first instruction in the transaction is
supposed to be able to read the *old* version of the account so it can
advance the nonce. but in this case, we're updating it before running
the instructions, and the instruction reads the *already updated*
version of the account, which leads to an unexpected error in the system
program.

the solution here is to copy the rollback account *before* modifying the
nonce. that way the account that will be read by the system program has
the expected old state, and the rollback account has the advanced nonce
(same as before, this is what we want). this is the same way that agave
works, and it's kind of confusing to identify this, but it all comes
down to the fact that Arc::make_mut is run in accountsdb, which clones
the underlying vec (not the arc) which means you get a copy of the
account.

now these are the steps
1. get pointer to nonce acct from cache
2. immediately copy that data and stop using the pointer to the cache
3. advance the nonce (mutating the copy)
4. use the copy as the rollback account

# Rent exemption check

The account should only get the rent exempt epoch if the account
***is*** exempt. The code was using `!= .Exempt` which is the opposite
of what we want. So I flipped the bool. Now it's also the same as agave.