fix(accountsdb): remove incorrect index code from stale commit in a feature branch
this was originally the planned way to add multiple index entries for the same account in the same slot. it was copied over from sebastian's branch. we decided not to take this approach and merged a different fix for updating an account in the same slot (not with multiple index entries)