fix(accountsdb): remove incorrect index code from stale commit in a feature branch (#830)
This code was copied over from sebastian's branch. It was originally how
we planned to add multiple index entries for the same account in the
same slot. We decided not to take this approach and merged a different
fix for updating an account in the same slot (not with multiple index
entries).

See this commit where I already removed this code
3fbe615b74b26a7e6b787f156f5dcbb961f16420