refactor(features): rework the feature set abstraction (#831)
A follow-up change will work on refactoring the feature set higher up,
storing it in the epoch constants rather than the slot constants. This
PR amends the call sites to ensure the slot is being passed in properly,
but it isn't necessary yet as the feature set is per-slot atm.

- Removes allocations from the feature set. There should never be a need
to do so, since the number of features is well known at compile time.
- Provides a single source of truth regarding the name of the feature,
its pubkey, and whether it has been activated on all clusters.
- Makes referencing the features more type-safe as well.