fix(runtime:program): Decode `ComputeBudgetInstruction` using BORSH codec (#834)
This type and relevant code interacting with it seem to have been
incorrectly ported without taking into account that it doesn't actually
use bincode, it uses [BORSH](https://borsh.io/).

The specific issue this causes is that the bincode decode process treats
tagged union tags as `u32`s, whereas BORSH treats them as `u8`s, which
was causing it to consume and interpret the 3 first bytes of the payload
(or whatever else comes after in the absence of a payload) as part of
the discriminant value.

In theory there's also other problems (that simply haven't and
*probably* won't manifest in practice [probably, hopefully]), like the
fact that where bincode permits encoding using either little or big
endian, BORSH only permits little endian, and where bincode has a
concept of varint encoding, BORSH always uses fixed-size int
representation.

In general, while there's a good degree of overlap between bincode and
BORSH, there's some small but important incompatibilities that make it
very risky to mix the two.

I'm open to suggestions about the BORSH abstraction, I implemented the
minimum possible amount of code whilst not entangling it too much with
`ComputeBudgetInstruction`'s definition and present assumptions.