refactor(ledger): rename blockstore to ledger (#859)
When I originally implemented the ledger, I left some confusing naming
inconsistencies in there, where "blockstore" is used sometimes and
"ledger" is used other times. I think we should consolidate on one name,
and my preference is "ledger."

This also fixes an issue in the ledger cleanup service where it would
sleep for 225 seconds, preventing shutdown. I fixed this as part of the
current pr because it was an easy way to increase the code coverage up
to the requirements.