feat(svm): add svm precompile entrypoints (#853)
Passed: 3373, Failed: 542 (86%)

Adds entrypoint for precompile invocation within the svm. Requires
instruction datas for whole transaction to be accessible from the
transaction context. This PR takes a simple approach of adding a
nullable instruction_datas slice to the transaction context which
populated if a precompile is present and the move_precompiles_to_svm
feature is active.