feat(replay): Initialize consensus state (#810)
Note: at the time of writing, there are changes included that remove the
RwMux wrappers from the slot and epoch trackers, and in general it's
currently presuming single-threaded execution, and this is largely for
simplicity. Later on when the vote listener PR stack is merged, I'll
devise a strategy for synchronizing access properly, with a full view of
the design, in such a way that it leaves open to us the possibility of
merging the thread of execution of the vote listener and replay, rather
than locking us into the architecture prescribed by agave.

---------

Co-authored-by: oluwadadepo aderemi <272535+dadepo@users.noreply.github.com>