fix(accountsdb): increase bincode allocation limit for snapshot deserialization (#883)
The default limit (100 MB) fails every time when loading testnet
snapshots. Even 1 GB is not enough memory. I bumped it up to 2 GB which
works.

It's a tight margin but I'd like to figure out *why* the manifest needs
so much memory before giving it too much leeway. If this fails in the
future, it will force us to think about it again, which is a good thing.
At this point we just need the minimal bugfix for this critical issue.