fix(replay): partial slot resume errors (#884)
Replay was failing to deserialize entries from shreds when resuming a
partially replayed slot because the shred starting point was off by one.
This fixes the ledger reader to return the correct number of shreds
processed.

After fixing that, replay fails to verify the PoH chain on resumed slots
because it didn't update the progress map to have the correct hash to
resume from. This fixes replay to set the correct hash in confirmation
progress.