fix(shred_network): Put upper limit on how many shreds to wait for (#896)
While debugging replay, we discovered this function ends up consuming
way too many shreds at a time, which causes parts of replay to wait for
unacceptably extended periods of time; 1024 is a best-guess estimate for
a balance between throughput and latency based on what we've observed to
be typical, which also doesn't cause the memory leak problem from
oversaturating the channel.

Also combines the two separately allocated arraylists into a single
multiarraylist.