fix(replay): error when feature account not owned by feature program (#912)
This should not result in an error. The account should simply not be
used to activate a feature unless it's owned by the feature program. See
the agave code in the permalink I added.

This is causing a problem now on testnet because the
`zkiTNuzBKxrCLMKehzuQeKZyLtX2yvFcEKMML8nExU8` account was recently
funded but ownership has not yet been transferred to the feature
program.

---------

Co-authored-by: Sebastian <15335529+Sobeston@users.noreply.github.com>