feat(replay, consensus): Add replay vote sender (#875)
This PR incorporates the replay vote sender, which is used to send vote
transaction in replay to vote listener.

It essentially those the following:

- Set up and start the vote listener in cmd.zig
- Set up `*sig.sync.Channel(ParsedVote)` in cmd.zig and passes to both
the vote_listeners (as receiver) and replay (as the sender)
- Add ability to parse ResolvedTransaction to ParsedVote
- In replay, after `processBatch` finds the vote transaction and sends
it through `*sig.sync.Channel(ParsedVote)` to the vote listener

The code to processed the parsed votes already exist in the
`listenAndConfirmVotes` method in the vote listener.

---------

Co-authored-by: Trevor Berrange Sanchez <trevor.berrange@syndica.io>