fix(tower): Set correct value for `threshold_depth`/`threshold_size` (#951)
The `canVoteOnCandidateSlot` check was returning `false` and hence the
slot the validator should vote: `vote_slot` was always set to null. This
was due to the wrong value of 0 being set for `threshold_depth` and
`threshold_size`

The `threshold_depth` indicates how far back in the tower’s vote history
to look when checking stake thresholds and `threshold_size` indicates
the minimum fraction of total stake that must have voted on the fork to
permit voting/switching.

So having the values set to 0 was wrong.

This PR fixes that by setting these values to the default just as it is
done in Agave.