fix(replay): memory and ci issues and add tests
this is a squash of a few commits from dnut/replay/simple before it was rebased then squashed.

that branch was first rebased to sign-off unsigned commits that someone pushed from an ec2 instance without signing. the changes in this PR were applied on top of the rebased branch after the sign-off.

the signoff didn't satisfy github for some reason, so i just went ahead and fully squashed it into a single commit. this was the easiest way to satisfy github, but it destroyed the history

this branch is meant to preserve the original history from before the initial rebase.

this commit just applies the diff from the few commits that occurred after the signoff, but before the squash, to put these changes top of the original unrebased version of that branch. the intent is to make it easier for people to integrate changes from dnut/replay/simple into their branches that are branched off that one