feat(replay): minimal and offline replay (#920)
This adds two features
- offline replay - replay a downloaded ledger
- minimal replay - run replay single-threaded without consensus

## CLI changes

- new subcommand: `replay-offline`: run replay offline, without gossip
or shred-network, replaying only pre-downloaded data.
- new option: `--disable-consensus`
- new option: `--replay-threads`

## Offline Replay

This mode runs replay without requiring any network connectivity. Gossip
and the shred network are not started. You must already have a snapshot
downloaded and some shreds in the ledger building off that snapshot.

## Single-threaded replay

Currently, replay is designed to be highly concurrent and asynchronous.
It executes batches concurrently based on account-locking logic, and it
executes slots concurrently when they fork.

This adds a variant of replay that executes entirely single-thread. The
concurrent version is having some memory safety issues. We just need
something that works and my hope is this can avoid the thread-safety
bugs that we haven't figured out yet.

## Disable consensus

When consensus is disabled, it runs `bypassConsensus`, which freezes the
slot and optimistically roots slots using SlotTree, which was added as a
simplified alternative to tracking forks with consensus. The basic idea
is that it roots slots after 32 slots.

## Diff Overview

Most of the diff here is refactoring existing code. This PR adds three
new run modes (online/simple, offline/full, offline/simple). To avoid
copy and pasting everything three times, I extracted out logic that is
used in multiple run modes. This also helped organization. For example,
there's a clear division between the consensus-specific state and the
other replay state, which are now distinguished with different structs.