refactor(replay): consolidate layers and separate consensus (#957)
Replay has a lot of layers of structs and files that make the code
confusing to navigate and make it difficult to pass state down through
the layers. This PR consolidates some of those layers. It also separates
the consensus code into its own folder to fully decouple it from replay
execution.