test: disable test that hangs on MacOS 26 (#974)
On macOS 26 this test just hangs; the `send`s are happening with data,
but the `recv`s never get anything.
why? No idea! This could be either a macOS kernel bug or a user bug,
although I don't see it being the latter yet, since the code is simple
and all seems correct enough.