ci(codecov): set patch requirement to 85% (#959)
Currently all PRs are required to have the lines in their diff covered
at least as well as main. So if main is at 91% coverage, 91% of the
lines in your diff need to be covered. The good thing about this
approach is that it guarantees the overall coverage is always
increasing. But 100% coverage is not practically achievable, so you need
to transition away from this approach at some point.

At 91% (current coverage on on main), the requirement is *usually* still
achievable, but the amount of effort required to reach that number is
problematic because we are starting to spend more time adding unhelpful
or irrelevant tests to satisfy codecov than we do implementing the
actual code that matters.

Also the quality of the tests diminishes as you are forced to write
contrived code just to hit every single line of code. Code coverage is
already a very poor metric that doesn't represent how well tested the
codebase is, which can lead to a false sense of security when the number
is high. These contrived tests only amplify the problem.

I set the number to 85%, which is an arbitrary number. I think even 85%
is probably still too high, but I want to be conservative and make sure
this change is actually an improvement.

As reviewers, we should consider the code coverage when we review the
code. The hard limit is there as a basic sanity check, but it's really
up to us as humans to understand whether the code is covered well, since
there is no tool that can answer that question for us.

---------

Co-authored-by: David Rubin <david@vortan.dev>
Co-authored-by: InKryption <59504965+InKryption@users.noreply.github.com>