fix(accountsdb): proper ordering/duration of rwlock acquisitions in putAccount (#948)
The original ordering of this function permitted a situation where a
lock would be acquired on the pubkey_ref_map, during a period in time in
which another function had acquired a lock on the slot_reference_map,
before it manages to acquire the lock on the pubkey_ref_map. The exact
semantics of where the edge case mutation during this race occurs
escapes me, but it isn't hard to see that having such scattered lock
ordering would lead to state inconsistencies, and this fixes that,
eliminating the invalid fee payer account error.