docs: list all conformance commits (#977)
This adds `conformance/commits.env` with a commit hash of every repo
used for testing sig's conformance with agave.

It's important that these commits are the same commits that we're using
in CI. Anyone should be able to easily check out our code and run
conformance locally by using these commit hashes, with high confidence
that everything will work and be conformant with the code we have on
`main`.

For SOLANA_CONFORMANCE_COMMIT and TEST_VECTORS_COMMIT, I changed the CI
to ensure these are always pulled in directly from commits.env and used
to check out the actual code used in CI.

For the other commits, this is not possible since the artifacts are
pre-generated and not checked out within CI. So I added a basic script
to grep files for where these commit hashes are documented or used for
other purposes, and set it up to run this verification script in CI. It
isn't a foolproof check, but at least it will flag some inconsistencies.
The [first
run](https://app.circleci.com/pipelines/github/Syndica/sig/5596/workflows/39e739e0-b817-44f9-a17f-33fe2ca02b34/jobs/83630)
detected an inconsistency that we have on main which I fixed in
[1434e9d](https://github.com/Syndica/sig/pull/977/commits/1434e9d05762ba11b9deb0821d22ce1322f5b7ec).