refactor(ledger): consolidate ledger state into Ledger struct (#995)
This refactors the ledger to consolidate its state into an easier to use
`Ledger` struct that can be used as the sole dependency to construct
`ledger.Reader`, `ledger.ResultWriter`, and `ledger.ShredInserter`.

While the diff appears huge, the change is just a reorganization of
pre-existing code. This refactoring allowed me to remove over 500 lines
of boilerplate code that were required with the old design. There are no
semantic changes except a few small bug fixes in `Reader` revealed by
some tests I added to satisfy codecov.

# Details

The ledger has three pieces of state: LedgerDB, max_root, and
highest_slot_cleaned. Previously they were separately initialized, and
then three separate pointers to each of these three items needed to be
separately passed into the three structs that use them: ShredInserter,
LedgerReader, and LedgerResultWriter to initialize them. Then those
three structs needed to be passed around to interact with the ledger.
This 3x3 approach is confusing and painful to manage/instantiate.

Now all the state (LedgerDB, max_root, and highest_slot_cleaned) is
neatly packaged into a new struct called `Ledger`. which is the only
dependency you need to pass around. It represents all of the ledger
state, and you can easily use it to insert shreds into the ledger, read
from the ledger, or write results to the ledger.

I removed UnifiedLedger since it was a half-baked approach at packaging
everything together. You can think of the new approach as a sort of
replacement for UnifiedLedger that is also intended to be used by any
components that interact with the ledger.

I considered deleting the three structs ShredInserter, LedgerReader, and
LedgerResultWriter, and just turning their methods into normal
functions. The annoyance with that approach is that you'd need to
manually pass in the Ledger, their specific logger, and their specific
metrics into each of those function. This makes the ergonomics just as
bad as it was before. So I decided to keep around structs with those
names, only to neatly package those three items.

But you don't need to worry about managing those structs separately from
the ledger. `Ledger` is all you need to get any of those structs. I
added helper methods to the ledger for using them:

```zig
const block = ledger.reader().getCompleteBlock(allocator, slot);
```