ci: improve test output to detect hanging tests (#1008)
Currently when our tests run in CI, there is no output indicating which
tests are running. This is a problem for many reasons, one of which is
that we can't tell which test is hanging when it gets stuck.

Running the tests as `zig-out/bin/test 2>&1 | cat` gives us output that
looks like this:

```
1/285 tests.test_0...OK
2/285 runtime.sysvar.stake_history.test.serialize and deserialize...OK
3/285 runtime.program.vote.state.test.AuthorizeVoters.serialize...OK
4/285 runtime.program.vote.state.test.Lockout.lockout...OK
5/285 runtime.program.vote.state.test.state.Lockout.lastLockedOutSlot...OK
6/285 runtime.program.vote.state.test.state.Lockout.isLockedOutAtSlot...OK
7/285 runtime.program.vote.state.test.state.VoteState.convertToCurrent...OK
```

Now if a test is hanging in CI, it will be clear when the job times out
which test is causing the issue.